refact: project structure
