+++ src/typecheck/model/props.ts
@@ -1,5 +1,5 @@
-import { Lens } from "./action";
 import { Term } from "./term";
+import {Lens} from "./lens";
 
 export interface TermProps<T extends Term> {
   term: T,
