+++ src/view/term/TermAny.tsx
@@ -1,12 +1,12 @@
-import { TAny } from "../typecheck/model/term";
-import { TermProps } from "../typecheck/model/props";
-import { Droppable } from "../component/Droppable";
-import { applyLens, joinLens } from "../typecheck/model/rec";
-import { Block } from "../component/Block";
+import { TAny } from "../../typecheck/model/term";
+import { TermProps } from "../../typecheck/model/props";
+import { Droppable } from "../../component/Droppable";
+import { applyLens, joinLens } from "../../typecheck/model/lens";
+import { Block } from "../../component/Block";
 import { useSnapshot } from "valtio";
-import { state } from "../state";
-import { prettyStep } from "../typecheck/pretty";
-import { quote } from "../typecheck/quote";
+import { state } from "../../state";
+import { prettyStep } from "../../typecheck/pretty";
+import { quote } from "../../typecheck/quote";
 
 export function TermAny({ lens, parent }: TermProps<TAny>): JSX.Element {
   const inferRes = useSnapshot(applyLens(state.inferResult, lens)) as any
