+++ src/view/term/TermType.tsx
@@ -1,11 +1,11 @@
 
-import { TType } from "../typecheck/model/term";
-import { TermProps } from "../typecheck/model/props";
-import { mkAction } from "../typecheck/model/action";
-import { onUpdate } from "../state";
-import { Block } from "../component/Block";
-import Input from "../component/Input";
-import { palette } from "./color";
+import { TType } from "../../typecheck/model/term";
+import { TermProps } from "../../typecheck/model/props";
+import { mkAction } from "../../typecheck/model/action";
+import { Block } from "../../component/Block";
+import Input from "../../component/Input";
+import { palette } from "../color";
+import {onUpdate} from "../../state/onUpdate";
 
 export function TermType({ term, lens, parent }: TermProps<TType>): JSX.Element {
   return <Block
